Refactored and added validation tests
