+++ rosetta-integration-tests/src/test/java/com/regnosys/rosetta/validation/RosettaValidatorTest.java
@@ -170,89 +170,6 @@ public class RosettaValidatorTest extends AbstractValidatorTest {
     }
 
     @Test
-    void testSwitchDefaultMustComeAtTheEnd() {
-        var expression = expressionParser.parseExpression("""
-				42 switch
-					default False,
-					42 then True
-				""");
-        validationTestHelper.assertError(expression, SWITCH_CASE_OR_DEFAULT, null, "A default case is only allowed at the end");
-    }
-
-    @Test
-    void testSwitchOnChoiceCannotHaveLiteralGuard() {
-        var model = modelHelper.parseRosettaWithNoIssues("""
-				choice Foo:
-					Opt1
-					Opt2
-
-				type Opt1:
-
-				type Opt2:
-				""");
-
-        var expression = expressionParser.parseExpression("""
-				foo switch
-					Opt1 then 1,
-					"Hello" then 2
-				""", List.of(model), List.of("foo Foo (1..1)"));
-        validationTestHelper.assertError(expression, SWITCH_CASE_OR_DEFAULT, null, "Case should match a choice option of type Foo");
-    }
-
-    @Test
-    void testSwitchOnChoiceWithMissingCase() {
-        var model = modelHelper.parseRosettaWithNoIssues("""
-				choice Foo:
-					Opt1
-					Bar
-
-				choice Bar:
-					Opt2
-					Opt3
-
-				type Opt1:
-
-				type Opt2:
-
-				type Opt3:
-				""");
-
-        var expression = expressionParser.parseExpression("""
-				foo switch
-					Opt1 then 1,
-					Opt2 then 2
-				""", List.of(model), List.of("foo Foo (1..1)"));
-        validationTestHelper.assertError(expression, SWITCH_OPERATION, null, "Missing the following cases: Opt3. Either provide all or add a default.");
-    }
-
-    @Test
-    void testSwitchWithNonReachableCase() {
-        var model = modelHelper.parseRosettaWithNoIssues("""
-				choice Foo:
-					Opt1
-					Bar
-
-				choice Bar:
-					Opt2
-					Opt3
-
-				type Opt1:
-
-				type Opt2:
-
-				type Opt3:
-				""");
-
-        var expression = expressionParser.parseExpression("""
-				foo switch
-					Opt1 then 1,
-					Bar then 2,
-					Opt2 then 3
-				""", List.of(model), List.of("foo Foo (1..1)"));
-        validationTestHelper.assertError(expression, SWITCH_CASE_OR_DEFAULT, null, "Case already covered by Bar");
-    }
-
-    @Test
     void testCanUseMixOfImportAliasAndFullyQualified() {
         var model1 = """
 				namespace foo.bar
@@ -332,152 +249,6 @@ public class RosettaValidatorTest extends AbstractValidatorTest {
         validationTestHelper.assertError(model, IMPORT, null,
                 "\"as\" statement can only be used with wildcard imports"
         );
-    }
-
-    @Test
-    void testSwitchWithDuplicateCase() {
-        var expression = expressionParser.parseExpression("""
-				42 switch
-					0 then 1,
-					1 then 2,
-					0 then 3
-				""");
-        validationTestHelper.assertError(expression, SWITCH_CASE_OR_DEFAULT, null, "Duplicate case");
-    }
-
-    @Test
-    void testSwitchInputRecordTypesAreNotValid() {
-        var expression = expressionParser.parseExpression("someDate switch default \"someResult\"", List.of("someDate date (1..1)"));
-        validationTestHelper.assertError(expression, SWITCH_OPERATION, null, "Operator `switch` is not supported for type `date`. Supported argument types are basic types, enumerations, complex types, and choice types");
-    }
-
-    @Test
-    void testSwitchWithMultiCardinalityInputIsInvalid() {
-        var model = modelHelper.parseRosettaWithNoIssues("""
-				enum SomeEnum:
-					A
-					B
-					C
-					D
-				""");
-
-        var expression = expressionParser.parseExpression("""
-				inEnum switch
-					A then "aValue",
-					B then "bValue",
-					C then "cValue",
-					default "someOtherValue"
-				""", List.of(model), List.of("inEnum SomeEnum (1..*)")); 
-        validationTestHelper.assertWarning(expression, ROSETTA_EXPRESSION, null, "Expecting single cardinality. The `switch` operator requires a single cardinality input");
-    }
-
-    @Test
-    void testValidSwitchSyntaxEnumIsValidWhenMissingEnumValuesWithDefault() {
-        var model = modelHelper.parseRosettaWithNoIssues("""
-				enum SomeEnum:
-					A
-					B
-					C
-					D
-				""");
-
-        var expression = expressionParser.parseExpression("""
-				inEnum switch
-					A then "aValue",
-					B then "bValue",
-					C then "cValue",
-					default "someOtherValue"
-				""", List.of(model), List.of("inEnum SomeEnum (1..1)"));
-        validationTestHelper.assertNoIssues(expression);
-    }
-
-    @Test
-    void testValidSwitchSyntaxEnumFailsWhenMissingEnumValues() {
-        var model = modelHelper.parseRosettaWithNoIssues("""
-				enum SomeEnum:
-					A
-					B
-					C
-					D
-				""");
-
-        var expression = expressionParser.parseExpression("""
-				inEnum switch
-					A then "aValue",
-					B then "bValue",
-					C then "cValue"
-				""", List.of(model), List.of("inEnum SomeEnum (1..1)"));
-        validationTestHelper.assertError(expression, SWITCH_OPERATION, null, "Missing the following cases: D. Either provide all or add a default.");
-    }
-
-    @Test
-    void testSwitchArgumentMatchesCaseStatementTypes() {
-        var model = modelHelper.parseRosettaWithNoIssues("""
-				enum SomeEnum:
-					A
-					B
-					C
-					D
-				""");
-
-        var expression = expressionParser.parseExpression("""
-				inEnum switch
-					A 	then "aValue",
-					10 	then "bValue",
-					C 	then "cValue",
-					default "defaultValue"
-				""", List.of(model), List.of("inEnum SomeEnum (1..1)"));
-        validationTestHelper.assertError(expression, SWITCH_CASE_OR_DEFAULT, null, "Case should match an enum value of SomeEnum");
-    }
-
-    @Test
-    void testValidSwitchSyntaxWithDefault() {
-        var context = modelHelper.parseRosettaWithNoIssues("""
-				enum SomeEnum:
-					A
-					B
-					C
-					D
-				""");
-
-        var expression = expressionParser.parseExpression("""
-				inEnum switch
-					A then "aValue",
-					B then "bValue",
-					C then "cValue",
-					default "defaultValue"
-				""", List.of(context), List.of("inEnum SomeEnum (1..1)"));
-        validationTestHelper.assertNoIssues(expression);
-    }
-
-    @Test
-    void testValidSwitchSyntaxEnum() {
-        var model = modelHelper.parseRosettaWithNoIssues("""
-				enum SomeEnum:
-					A
-					B
-					C
-					D
-				""");
-
-        var expression = expressionParser.parseExpression("""
-				inEnum switch
-					A then "aValue",
-					B then "bValue",
-					C then "cValue",
-					D then "dValue"
-				""", List.of(model), List.of("inEnum SomeEnum (1..1)"));
-        validationTestHelper.assertNoIssues(expression);
-    }
-
-    @Test
-    void testValidSwitchSyntaxString() {
-        var expression = expressionParser.parseExpression("""
-				someInput switch
-					"A" then "aValue",
-					"B" then "bValue"
-				""", List.of("someInput string (1..1)"));
-        validationTestHelper.assertNoIssues(expression);
     }
 
     @Test
